extrai funções de BigDecimal
